+++ src/app/(with-layout)/page.js
@@ -9,8 +9,8 @@ import MarqueeCard from "../HomePage/MarqueeCard";
 import OfferService from "../HomePage/OfferService";
 import OtherSection from "../HomePage/OtherSection";
 // import SliderOffer from "../HomePage/SliderOffer";
+// import TabSection1 from "../HomePage/TabSection1";
 import TAbSection from "../HomePage/TAbSection";
-import TabSection1 from "../HomePage/TabSection1";
 import TradeTypes from "../HomePage/TradeTypes";
 import UpCommingOther from "../HomePage/UpCommingOther";
 import AboutPage from "./about/page";
