+++ src/app/dashboard/activities/Activities.js
