+++ src/app/dashboard/exchange/TablePriceDeep.js
