+++ src/app/dashboard/exchange/chartSection.js
@@ -14,21 +14,19 @@ const ChartSection = () => {
     <div className="mx-auto">
       <div className="flex gap-3  font-semibold mb-4">
         <button
-          className={`cursor-pointer ${
-            selectedChart === "price"
-              ? " btn btn-success btn-sm "
-              : "btn btn-sm btn-success btn-outline"
-          } mr-4`}
+          className={`cursor-pointer ${selectedChart === "price"
+            ? " btn btn-success btn-sm "
+            : "btn btn-sm btn-success btn-outline"
+            } mr-4`}
           onClick={() => handleSelectedChart("price")}
         >
           Price Chart
         </button>
         <button
-          className={`cursor-pointer ${
-            selectedChart === "deep"
-              ? " btn btn-sm btn-success"
-              : "btn btn-sm btn-success btn-outline"
-          } mr-4`}
+          className={`cursor-pointer ${selectedChart === "deep"
+            ? " btn btn-sm btn-success"
+            : "btn btn-sm btn-success btn-outline"
+            } mr-4`}
           onClick={() => handleSelectedChart("deep")}
         >
           Deep Chart
