+++ src/app/dashboard/layout.js
@@ -6,26 +6,26 @@ export const metadata = {
   description: 'Generated by create next app',
 }
 
-const DashboardLayout = ({children}) => {
+const DashboardLayout = ({ children }) => {
   return (
     <div className='flex bg-[#131720] text-slate-300 
     '>
       <div>
-      <Sidebar></Sidebar>
+        <Sidebar></Sidebar>
       </div>
 
       <div className='flex-1'>
         <DashboardNav></DashboardNav>
-      {children}
-      {/* footer */}
+        {children}
+        {/* footer */}
         <footer className="footer md:flex justify-between  py-4 px-7 bg-base-300 text-base-content">
-         
-            <p>Copyright © 2023. All Rights Reserved By OTAB</p>
-            <div className='flex gap-5 justify-between'>
-              <p>Help Center</p>
-              <p>Privacy</p>
-              <p>Terms of Service</p>
-            </div>
+
+          <p>Copyright © 2023. All Rights Reserved By OTAB</p>
+          <div className='flex gap-5 justify-between'>
+            <p>Help Center</p>
+            <p>Privacy</p>
+            <p>Terms of Service</p>
+          </div>
         </footer>
       </div>
     </div>
