+++ src/app/dashboard/prices/page.js
@@ -5,12 +5,12 @@ import Prices from "./Prices";
 export const metadata = {
   title: 'Prices | OTAB',
 }
- 
+
 
 const PricePage = () => {
   return (
-    <div> 
-      <Prices></Prices> 
+    <div>
+      <Prices></Prices>
     </div>
   );
 };
