+++ src/app/dashboard/wallets/page.js
@@ -1,11 +1,11 @@
-  
+
 import Wallets from './Wallets';
 
 
 const WalletsPage = () => {
   return (
     <div className='w-full'>
-         <Wallets></Wallets>
+      <Wallets></Wallets>
     </div>
   );
 };
